requirements-python.txt
# Python 카카오톡 YouTube 봇을 위한 의존성
pyautokakao>=0.1.0
pyautogui>=0.9.54
opencv-python>=4.8.0
pillow>=10.0.0
apify-client>=1.7.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
pyperclip>=1.8.2 